cloud-pubsub/requirements.txt
google-cloud-pubsub~=2.16.0
